Move away withSessionProtection from authenticated-container
