+++ packages/web/app/pages/[orgId]/[projectId]/[targetId]/explorer/[typename].tsx
@@ -1,7 +1,8 @@
 import { ReactElement } from 'react';
 import { DocumentType, gql, useQuery } from 'urql';
 
-import { authenticated, withSessionProtection } from '@/components/authenticated-container';
+import { authenticated } from '@/components/authenticated-container';
+import { withSessionProtection } from '@/lib/supertokens/guard';
 import { TargetLayout } from '@/components/layouts';
 import {
   GraphQLEnumTypeComponent,
