+++ packages/web/app/pages/[orgId]/[projectId]/[targetId]/history.tsx
@@ -5,7 +5,8 @@ import { VscBug, VscDiff, VscListFlat } from 'react-icons/vsc';
 import reactStringReplace from 'react-string-replace';
 import { useQuery } from 'urql';
 
-import { authenticated, withSessionProtection } from '@/components/authenticated-container';
+import { authenticated } from '@/components/authenticated-container';
+import { withSessionProtection } from '@/lib/supertokens/guard';
 import { Label } from '@/components/common';
 import { TargetLayout } from '@/components/layouts';
 import {
