+++ packages/web/app/pages/[orgId]/[projectId]/[targetId]/history/[versionId].tsx
@@ -1,4 +1,4 @@
-import { withSessionProtection } from '@/components/authenticated-container';
+import { withSessionProtection } from '@/lib/supertokens/guard';
 import HistoryPage from '../history';
 
 export const getServerSideProps = withSessionProtection();
