+++ packages/web/app/src/lib/api/extract-access-token-from-request.ts
@@ -3,7 +3,7 @@ import { superTokensNextWrapper } from 'supertokens-node/nextjs';
 import { verifySession } from 'supertokens-node/recipe/session/framework/express';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { backendConfig } from '@/config/supertokens/backend';
-import { SessionContainerInterface } from 'supertokens-node/lib/build/recipe/session/types';
+import type { SessionContainerInterface } from 'supertokens-node/lib/build/recipe/session/types';
 
 supertokens.init(backendConfig());
